feat: migrate snapshot to new plugin infra
