+++ packages/maskbook/src/plugin-infra/register.ts
@@ -3,4 +3,5 @@
 
 import '@dimensiondev/plugin-example'
 import '../plugins/RedPacket'
+import '../plugins/Snapshot'
 import '../plugins/External'
