+++ packages/maskbook/src/plugins/Collectible/UI/Collectible.tsx
@@ -33,7 +33,7 @@ import { currentCollectibleProviderSettings } from '../settings'
 import { FootnoteMenu, FootnoteMenuOption } from '../../Trader/UI/trader/FootnoteMenu'
 import { MaskbookTextIcon } from '../../../resources/MaskbookIcon'
 import { resolveAssetLinkOnOpenSea, resolveCollectibleProviderName } from '../pipes'
-import { Markdown } from '../../Snapshot/UI/Markdown'
+import { Markdown } from '../../Snapshot/SNSAdaptor/Markdown'
 import { ActionBar } from './ActionBar'
 import { useChainId } from '@dimensiondev/web3-shared'
 import { getEnumAsArray } from '@dimensiondev/maskbook-shared'
