+++ packages/maskbook/src/plugins/PluginUI.ts
@@ -21,7 +21,6 @@ import { AirdropPluginDefine } from './Airdrop/define'
 import { CollectiblesPluginDefine } from './Collectible/define'
 import { sideEffect } from '../utils/side-effects'
 import { VCentPluginDefine } from './VCent/define'
-import { SnapShotPluginDefine } from './Snapshot/define'
 import { DHedgePluginDefine } from './dHEDGE/define'
 
 sideEffect.then(() => {
@@ -36,7 +35,6 @@ sideEffect.then(() => {
     if (Flags.trader_enabled) plugins.add(TraderPluginDefine)
     if (Flags.transak_enabled) plugins.add(TransakPluginDefine)
     if (Flags.airdrop_enabled) plugins.add(AirdropPluginDefine)
-    if (Flags.snapshot_enabled) plugins.add(SnapShotPluginDefine)
     if (Flags.collectibles_enabled) plugins.add(CollectiblesPluginDefine)
     if (Flags.dhedge_enabled) plugins.add(DHedgePluginDefine)
     if (process.env.STORYBOOK) plugins.add(StorybookPluginDefine)
