+++ packages/maskbook/src/plugins/RedPacket/messages.ts
@@ -1,13 +1,15 @@
 import { RedPacketPluginID } from './constants'
 import { createPluginMessage } from '../utils/createPluginMessage'
 import { createPluginRPC } from '../utils/createPluginRPC'
+import type { WebExtensionMessage } from '@dimensiondev/holoflows-kit'
 
 export interface RedPacketMessages {
     redPacketUpdated: void
     _: unknown
 }
 if (import.meta.webpackHot) import.meta.webpackHot.accept()
-export const RedPacketMessage = createPluginMessage<RedPacketMessages>(RedPacketPluginID)
+export const RedPacketMessage: WebExtensionMessage<RedPacketMessages> =
+    createPluginMessage<RedPacketMessages>(RedPacketPluginID)
 export const RedPacketRPC = createPluginRPC(
     RedPacketPluginID,
     () => import('./Worker/services'),
