+++ packages/maskbook/src/plugins/Snapshot/SNSAdaptor/ReadmeCard.tsx
@@ -1,7 +1,7 @@
 import { useContext } from 'react'
 import { makeStyles } from '@material-ui/core'
 import { Markdown } from './Markdown'
-import { useProposal } from '../hooks/useProposal'
+import { useProposal } from './hooks/useProposal'
 import { SnapshotContext } from '../context'
 
 const useStyles = makeStyles((theme) => {
