+++ packages/maskbook/src/plugins/Snapshot/SNSAdaptor/ResultCard.tsx
@@ -4,12 +4,12 @@ import { makeStyles, Box, List, ListItem, Typography, LinearProgress, withStyles
 import { ShadowRootTooltip, useI18N } from '../../../utils'
 import millify from 'millify'
 import { SnapshotContext } from '../context'
-import { useProposal } from '../hooks/useProposal'
-import { useVotes } from '../hooks/useVotes'
-import { useResults } from '../hooks/useResults'
+import { useProposal } from './hooks/useProposal'
+import { useVotes } from './hooks/useVotes'
+import { useResults } from './hooks/useResults'
 import { SnapshotCard } from './SnapshotCard'
 import { parse } from 'json2csv'
-import { useRetry } from '../hooks/useRetry'
+import { useRetry } from './hooks/useRetry'
 import { LoadingFailCard } from './LoadingFailCard'
 import { LoadingCard } from './LoadingCard'
