+++ packages/maskbook/src/plugins/Snapshot/SNSAdaptor/VoteConfirmDialog.tsx
