+++ packages/maskbook/src/plugins/Snapshot/SNSAdaptor/VotesCard.tsx
@@ -7,8 +7,8 @@ import { useContext } from 'react'
 import { useI18N } from '../../../utils'
 import { EthereumBlockie } from '../../../web3/UI/EthereumBlockie'
 import { SnapshotContext } from '../context'
-import { useRetry } from '../hooks/useRetry'
-import { useVotes } from '../hooks/useVotes'
+import { useRetry } from './hooks/useRetry'
+import { useVotes } from './hooks/useVotes'
 import type { VoteItem } from '../types'
 import { LoadingCard } from './LoadingCard'
 import { LoadingFailCard } from './LoadingFailCard'
