+++ packages/maskbook/src/plugins/Snapshot/SNSAdaptor/helpers.ts
@@ -1,4 +1,4 @@
-import type { ProposalIdentifier } from './types'
+import type { ProposalIdentifier } from '../types'
 
 export function getProposalIdentifier(url: string): ProposalIdentifier {
     const parsedURL = new URL(url)
