+++ packages/maskbook/src/plugins/Snapshot/messages.ts
@@ -22,6 +22,6 @@ export const PluginSnapshotMessages: WebExtensionMessage<SnapshotMessages> =
     createPluginMessage<SnapshotMessages>(SNAPSHOT_PLUGIN_ID)
 export const PluginSnapshotRPC = createPluginRPC(
     SNAPSHOT_PLUGIN_ID,
-    () => import('./services'),
+    () => import('./Worker/services'),
     PluginSnapshotMessages.events.rpc,
 )
